backend/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
celery==5.3.4
redis==5.0.0
minio==7.1.16
python-dotenv==1.0.0
opencv-python-headless==4.7.0.72
numpy==1.26.0
Pillow==10.0.0
pyzbar==0.1.9
pytesseract==0.3.10
twilio==8.5.0
onnxruntime==1.19.2
requests==2.32.3
